Use new plugin format for existing plugins
