+++ buildSrc/src/main/kotlin/org/jetbrains/jewel/buildlogic/theme/IntelliJThemeGeneratorPlugin.kt
@@ -1,64 +1,20 @@
 package org.jetbrains.jewel.buildlogic.theme
 
 import com.squareup.kotlinpoet.ClassName
-import io.gitlab.arturbosch.detekt.Detekt
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.decodeFromStream
 import org.gradle.api.DefaultTask
 import org.gradle.api.NamedDomainObjectContainer
-import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
-import org.gradle.kotlin.dsl.container
-import org.gradle.kotlin.dsl.get
-import org.gradle.kotlin.dsl.getByType
 import org.gradle.kotlin.dsl.property
-import org.gradle.kotlin.dsl.register
-import org.gradle.kotlin.dsl.withType
-import org.gradle.util.internal.GUtil
-import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
-import org.jetbrains.kotlin.gradle.tasks.BaseKotlinCompile
 import java.net.URL
-
-abstract class IntelliJThemeGeneratorPlugin : Plugin<Project> {
-
-    final override fun apply(target: Project) {
-        with(target) {
-            val extension = ThemeGeneratorContainer(container<ThemeGeneration> { ThemeGeneration(it, project) })
-            extensions.add("intelliJThemeGenerator", extension)
-
-            extension.all {
-                val task = tasks.register<IntelliJThemeGeneratorTask>("generate${GUtil.toCamelCase(name)}Theme") {
-                    outputFile.set(targetDir.file(this@all.themeClassName.map {
-                        val className = ClassName.bestGuess(it)
-                        className.packageName.replace(".", "/")
-                            .plus("/${className.simpleName}.kt")
-                    }))
-                    themeClassName.set(this@all.themeClassName)
-                    ideaVersion.set(this@all.ideaVersion)
-                    themeFile.set(this@all.themeFile)
-                }
-                tasks.withType<BaseKotlinCompile> {
-                    dependsOn(task)
-                }
-                tasks.withType<Detekt> {
-                    dependsOn(task)
-                }
-                pluginManager.withPlugin("org.jetbrains.kotlin.jvm") {
-                    extensions.getByType<KotlinJvmProjectExtension>().apply {
-                        sourceSets["main"].kotlin.srcDir(targetDir)
-                    }
-                }
-            }
-        }
-    }
-}
 
 class ThemeGeneratorContainer(container: NamedDomainObjectContainer<ThemeGeneration>) : NamedDomainObjectContainer<ThemeGeneration> by container
